# doc/api.yaml
openapi: 3.0.0
info:
    title: Home Library Service
    description: Home music library service
    version: 1.0.0
servers:
    - url: /
components:
    schemas:
        User:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                login:
                    type: string
                    example: 'TestUser'
                version:
                    type: number
                    example: 1
                createAt:
                    type: number
                    example: 1655000000
                updateAt:
                    type: number
                    example: 1655000000
            required:
                - id
                - login
...
        Track:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                    example: The Show Must Go On
                artistId:
                    type: string
                    format: uuid
                    nullable: true
                albumId:
                    type: string
                    format: uuid
                    nullable: true
                duration:
                    type: integer
                    description: In seconds
                    example: 262
            required:
                - id
                - name
                - duration
...
    responses:
        UnauthorizedError:
            description: Access token is missing or invalid
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
security:
    - bearerAuth: []
paths:
    /user:
        get:
            tags:
                - Users
            summary: Get all users
            description: Gets all users
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/User'
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
        post:
            tags:
                - Users
            summary: Create user
            description: Creates a new user
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                login:
                                    type: string
                                    description: The user's login
                                password:
                                    type: string
                                    description: The user's password
                            required:
                                - login
                                - password
            responses:
                201:
                    description: The user has been created.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                400:
                    description: Bad request. body does not contain required fields
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
    /user/{userId}:
        parameters:
            - name: userId
              in: path
              required: true
              schema:
                  type: string
        get:
            tags:
                - Users
            summary: Get single user by id
            description: Get single user by id
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                400:
                    description: Bad request. Validation failed (uuid  is expected)
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: User not found

        put:
            tags:
                - Users
            summary: Update a user's password
            description: Updates a user's password by ID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                oldPassword:
                                    type: string
                                    description: The user's old password
                                newPassword:
                                    type: string
                                    description: The user's new password
                            required:
                                - oldPassword
                                - newPassword
            responses:
                200:
                    description: The user has been updated.
                    content:
                        application/json:
                            schema:
                                type: object
                                title: example
                                properties:
                                    id:
                                        type: string
                                        format: uuid
                                    login:
                                        type: string
                                        example: TestUser
                                    version:
                                        type: integer
                                        example: 2
                                    createAt:
                                        type: integer
                                        example: 1655000000
                                    updateAt:
                                        type: integer
                                        example: 1655999999
                400:
                    description: Bad request. Validation failed (uuid  is expected)
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
                403:
                    description: oldPassowrd is wrong
                404:
                    description: User not found

        delete:
            tags:
                - Users
            summary: Delete user
            description: Deletes user by ID.
            responses:
                204:
                    description: The user has been deleted
                400:
                    description: Bad request. userId is invalid (not uuid)
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: User not found
    /track:
        get:
            tags:
                - Track
            summary: Get tracks list
            description: Gets all library tracks list
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Track'
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
        post:
            tags:
                - Track
            summary: Add new track
            description: Add new track information
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                artistId:
                                    type: string
                                    format: uuid
                                albumId:
                                    type: string
                                    format: uuid
                                duration:
                                    type: integer
                                    description: In seconds
                            required:
                                - name
                                - duration
            responses:
                201:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Track'
                400:
                    description: Bad request.
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
    /track/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags:
                - Track
            summary: Get single track by id
            description: Gets single track by id
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Track'
                400:
                    description: Bad request. Validation failed (uuid  is expected)
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Track was not found.
        put:
            tags:
                - Track
            summary: Update track information
            description: Update library track information by UUID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            title: example
                            properties:
                                name:
                                    type: string
                                    example: Bohemian Rhapsody
                                artistId:
                                    type: string
                                    format: uuid
                                    nullable: true
                                duration:
                                    type: integer
                                    description: In seconds
                                    example: 355
                                albumId:
                                    type: string
                                    format: uuid
                                    nullable: true
                            required:
                                - name
                                - duration
            responses:
                200:
                    description: The track has been updated.
                    content:
                        application/json:
                            schema:
                                type: object
                                title: example
                                properties:
                                    id:
                                        type: string
                                        format: uuid
                                    name:
                                        type: string
                                        example: Bohemian Rhapsody
                                    artistId:
                                        type: string
                                        format: uuid
                                        nullable: true
                                    duration:
                                        type: integer
                                        description: In seconds
                                        example: 355
                                    albumId:
                                        type: string
                                        format: uuid
                                        nullable: true
                400:
                    description: Bad request. Validation failed (uuid  is expected)
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Track was not found.
        delete:
            tags:
                - Track
            summary: Delete track
            description: Delete track from library
            responses:
                204:
                    description: Delelted succesfully
                400:
                    description: Bad request. Validation failed (uuid  is expected)
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Track was not found.
...
    /favs/track/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: string
                  format: uuid
        post:
            tags:
                - Favorites
            summary: Add track to the favorites
            description: Add track to the favorites
            responses:
                201:
                    description: Added succesfully
                400:
                    description: Bad Request. Validation failed (uuid  is expected)
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
                422:
                    description: Track with id {trackId} does not exist
        delete:
            tags:
                - Favorites
            summary: Delete track from favorites
            description: Delete track from favorites
            responses:
                204:
                    description: Deleted succesfully
                400:
                    description: Bad Request. Validation failed (uuid  is expected)
                # 401:
                #   $ref: '#/components/responses/UnauthorizedError'
                404:
                    description: Track was not found.
...